requirements.txt
pdf2image
streamlit
google-generativeai
python-dotenv